manager-panel/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-dateutil==2.8.2

